+++ app/frontend/src/javascript/components/authentication-provider/data-mapping-form.tsx
@@ -114,7 +114,7 @@ export const DataMappingForm = <TFieldValues extends FieldValues, TContext exten
    * Return a className based on the current mapping-item status
    */
   const itemStatus = (index: number): string => {
-    if (currentFormValues[index]?.id) {
+    if (currentFormValues && currentFormValues[index]?.id) {
       if (currentFormValues[index]._destroy) return 'destroyed-item';
       return 'saved-item';
     }